user-service/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.1
flask-cors==3.0.10  # Add this line for CORS support
